Insurance Microservices/Claims/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
httpx
cryptography